+++ source/fortran/global_variables.f90
@@ -174,8 +174,8 @@ module physics_variables
 
   public
 
-  !+ad_vars  ipnlaws /46/ FIX : number of energy confinement time scaling laws
-  integer, parameter :: ipnlaws = 46
+  !+ad_vars  ipnlaws /47/ FIX : number of energy confinement time scaling laws
+  integer, parameter :: ipnlaws = 47
 
   !+ad_vars  abeam : beam ion mass (amu)
   real(kind(1.0D0)) :: abeam = 0.0D0
@@ -557,8 +557,10 @@ module physics_variables
        'Hubbard 2017 - lower (I)', &
   !+ad_varc  <LI> (45)  Hubbard 2017 (I-mode) - upper bound
        'Hubbard 2017 - upper (I)', &
-  !+ad_varc  <LI> (46)  NSTX (H-mode; Spherical tokamak)</UL>
-       'NSTX (Spherical)     (H)' /)
+  !+ad_varc  <LI> (46)  NSTX (H-mode; Spherical tokamak)
+       'NSTX (Spherical)     (H)', &
+  !+ad_varc  <LI> (47)  NSTX-Petty08 Hybrid (H-mode)</UL>
+       'NSTX-Petty08 Hybrid  (H)' /)
 
   !+ad_vars  iscrp /1/ : switch for plasma-first wall clearances:<UL>
   !+ad_varc         <LI> = 0 use 10% of rminor;