+++ CHANGELOG.md
@@ -12,7 +12,7 @@
   paper or user defined inputs). See Userguide for more info.
 - Updated version of Kallenbach scanning (now can specify the variable to scan 
   from a selection, number of scan points etc.). See Userguide for more info.
-- Added NSTX confinement time scaling (#820)
+- Added NSTX and NSTX-Petty08 confinement time scalings (#820)
 
 ## Minor Changes